Part two of the puzzle was a bit of a headscratcher.
The issue was replacing the str representatons of the numbers when the
numbers were daisy chained (i.e sevenine). We needed to replace with the string
injecting the number somewhere in the middle.

Then one pass didn't quite cut it because of how regex matches I guess. So
threw in another pass for good measure and we got the answer.

This is not efficient imo.